bot/requirements.txt
# Основные зависимости
python-telegram-bot>=22.1  # Последняя стабильная версия
python-dotenv>=1.1.0  # Для работы с переменными окружения
pydantic>=2.11.7  # Для валидации данных
fuzzywuzzy>=0.18.0  # Для нечеткого поиска
python-Levenshtein>=0.23.0  # Для оптимизации поиска
nest-asyncio>=1.5.8  # Для исправления event loop на Windows

# Зависимости для разработки
pytest>=8.4.1  # Для тестирования
pytest-cov>=6.2.1  # Для покрытия тестами
black>=25.1.0  # Для форматирования кода
isort>=6.0.1  # Для сортировки импортов
mypy>=1.16.1  # Для проверки типов
flake8>=7.2.0  # Для линтинга

# Зависимости для тестирования
pytest-asyncio>=1.0.0  # Для асинхронного тестирования
pytest-mock>=3.14.1  # Для мокирования
pytest-xdist>=3.5.0  # Для параллельного тестирования
coverage>=7.9.1  # Для анализа покрытия

# Зависимости для документации
sphinx>=8.2.3  # Для генерации документации
sphinx-rtd-theme>=3.0.2  # Тема документации
sphinx-autodoc-typehints>=1.25.2  # Для типов в документации

# Зависимости для логирования
loguru>=0.7.3  # Для улучшенного логирования

# Зависимости для безопасности
cryptography>=45.0.4  # Для шифрования
python-jose>=3.5.0  # Для JWT
bandit>=1.7.5  # Для проверки безопасности
safety>=2.3.5  # Для проверки уязвимостей

# Зависимости для оптимизации
ujson>=5.10.0  # Для быстрой работы с JSON
orjson>=3.10.18  # Для быстрой сериализации

# HTTP клиент
aiohttp>=3.12.13  # Для асинхронных HTTP запросов

# Мониторинг
prometheus-client>=0.19.0  # Для метрик
psutil>=5.9.6  # Для системных метрик

# Утилиты
python-dateutil>=2.9.0.post0  # Для работы с датами
pytz>=2025.2  # Для работы с часовыми поясами
requests>=2.32.4  # Для HTTP запросов

# Конфигурация
pydantic-settings>=2.1.0  # Для управления настройками

# Инструменты разработки
ipython>=8.21.0  # Для интерактивной разработки
ipdb>=0.13.13  # Для отладки
watchdog>=3.0.0  # Для отслеживания изменений файлов

# Кэширование
hiredis>=2.3.2  # Для работы с Redis

# Парсинг
beautifulsoup4>=4.12.2  # Для парсинга HTML

# Git хуки
pre-commit>=3.5.0  # Для проверки перед коммитом

# Утилиты
tqdm>=4.66.1  # Для прогресс-баров
rich>=13.6.0  # Для красивого вывода
typer>=0.9.0  # Для CLI
click>=8.2.1  # Для CLI
colorama>=0.4.6  # Для цветного вывода
